part1-database-etl/requirements.txt
pandas>=2.0.0
mysql-connector-python>=9.0.0
python-dotenv>=1.0.0
